// packages/app/src/routes/URLS.d.ts
export declare const URLS: {
    LOGIN: {
        (options?: {
            params?: Record<string, unknown> | undefined;
            query?: string | undefined;
            hash?: string | undefined;
            fragment?: string | undefined;
        }): string;
        pattern: string;
    };
    MAIL_INVITATION_SIGNUP: {
        (options?: {
            params?: Record<string, unknown> | undefined;
            query?: string | undefined;
            hash?: string | undefined;
            fragment?: string | undefined;
        }): string;
        pattern: string;
    };
    INVALID_MAIL_INVITATION_SIGNUP_TOKEN: {
        (options?: {
            params?: Record<string, unknown> | undefined;
            query?: string | undefined;
            hash?: string | undefined;
            fragment?: string | undefined;
        }): string;
        pattern: string;
    };
    RESET_PASSWORD: {
        (options?: {
            params?: Record<string, unknown> | undefined;
            query?: string | undefined;
            hash?: string | undefined;
            fragment?: string | undefined;
        }): string;
        pattern: string;
    };
    SET_PASSWORD: {
        (options?: {
            params?: Record<string, unknown> | undefined;
            query?: string | undefined;
            hash?: string | undefined;
            fragment?: string | undefined;
        }): string;
        pattern: string;
    };
    DASHBOARD: {
        (options?: {
            params?: Record<string, unknown> | undefined;
            query?: string | undefined;
            hash?: string | undefined;
            fragment?: string | undefined;
        }): string;
        pattern: string;
    };
};